requirements.txt
asgiref==3.8.1
dj-database-url==3.0.0
Django==5.2.3
django-cors-headers==4.7.0
django-decouple==2.1
django-filter==25.1
django-sendgrid-v5==1.3.0
djangorestframework==3.16.0
ecdsa==0.19.1
gunicorn==23.0.0
MarkupSafe==3.0.2
numpy==2.3.1
packaging==25.0
pandas==2.3.1
psycopg2-binary==2.9.10
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
pytz==2025.2
sendgrid==6.12.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.0
tzdata==2025.2
Werkzeug==3.1.3
whitenoise==6.9.0
